requirements.txt
pytest==7.1.2
allure-pytest==2.13.2
selene==2.0.0rc4